nolibc: avoid unnecessary padding in data structures

+++ tools/include/nolibc/std.h
@@ -25,7 +25,7 @@ typedef unsigned int         mode_t;
 typedef   signed int          pid_t;
 typedef unsigned int          uid_t;
 typedef unsigned int          gid_t;
-typedef unsigned long       nlink_t;
+typedef unsigned int        nlink_t;
 typedef  int64_t              off_t;
 typedef   signed long     blksize_t;
 typedef   signed long      blkcnt_t;
